refactor(utils): create new utils package and transfer some functions to marketplace.py
