+++ utils/items.py
@@ -1,0 +1,3 @@
+"""
+Item state management and database operations utilities.
+"""
