+++ utils/notifications.py
@@ -1,0 +1,3 @@
+"""
+User feedback system utilities for displaying messages and notifications.
+"""
