+++ utils/price_display.py
@@ -1,0 +1,3 @@
+"""
+Price-related UI enhancement utilities for displaying and formatting price data.
+"""
